Use thread-safe wrapper for ElasticSearchRestDAOV6 bulk request
