+++ es5-persistence/src/main/java/com/netflix/conductor/dao/es5/index/ElasticSearchRestDAOV5.java
@@ -47,7 +47,6 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.TimeZone;
-import java.util.Objects;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ExecutorService;
@@ -57,7 +56,6 @@ import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
-import javax.annotation.Nonnull;
 import javax.annotation.PreDestroy;
 import javax.inject.Inject;
 import javax.inject.Singleton;
@@ -951,41 +949,6 @@ public class ElasticSearchRestDAOV5 implements IndexDAO {
         BulkRequests(long lastFlushTime, BulkRequest bulkRequestWrapper) {
             this.lastFlushTime = lastFlushTime;
             this.bulkRequestWrapper = new BulkRequestWrapper(bulkRequestWrapper);
-        }
-
-        /**
-         * Thread-safe wrapper for {@link BulkRequest}.
-         */
-        private static class BulkRequestWrapper
-        {
-            private final BulkRequest bulkRequest;
-
-            BulkRequestWrapper(@Nonnull BulkRequest bulkRequest) {
-                this.bulkRequest = Objects.requireNonNull(bulkRequest);
-            }
-
-            public void add(@Nonnull UpdateRequest req) {
-                synchronized (bulkRequest) {
-                    bulkRequest.add(Objects.requireNonNull(req));
-                }
-            }
-
-            public void add(@Nonnull IndexRequest req) {
-                synchronized (bulkRequest) {
-                    bulkRequest.add(Objects.requireNonNull(req));
-                }
-            }
-
-            private BulkRequest get()
-            {
-                return bulkRequest;
-            }
-
-            int numberOfActions() {
-                synchronized (bulkRequest) {
-                    return bulkRequest.numberOfActions();
-                }
-            }
         }
     }
 }
